+++ tests/testthat/test-envvar.R
@@ -1,3 +1,4 @@
+
 test_that("can add before or after", {
   path_string <- function(...) {
     paste(c(...), collapse = .Platform$path.sep)