+++ NEWS.md
@@ -1,6 +1,9 @@
 
 # renv 1.1.0  (UNRELEASED)
 
+* Fixed an issue where `RENV_CONFIG_EXTERNAL_LIBRARIES` was incorrectly
+  split when using Windows paths. (#2069)
+  
 * Fixed an issue where `renv` failed to restore packages installed from
   r-universe when the associated lockfile record had no `RemoteRef` field.
   (#2068)